movies_admin/requirements/base.txt
asgiref==3.4.1
Django==3.2.0
django-environ==0.4.5
gunicorn==20.0.4
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
